+++ packages/db-collection-e2e/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @tanstack/db-collection-e2e
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [[`a90f95e`](https://github.com/TanStack/db/commit/a90f95eaeb7cc334955ae5e8ffb98940fce1ecf1), [`846a830`](https://github.com/TanStack/db/commit/846a8309a243197245f4400a5d53cef5cec6d5d9), [`8e26dcf`](https://github.com/TanStack/db/commit/8e26dcfde600e4a18cd51fbe524560d60ab98d70)]:
+  - @tanstack/query-db-collection@1.0.2
+  - @tanstack/db@0.5.3
+  - @tanstack/electric-db-collection@0.2.3
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [[`99a3716`](https://github.com/TanStack/db/commit/99a371630b9f4632db86c43357c64701ecb53b0e)]:
+  - @tanstack/db@0.5.2
+  - @tanstack/electric-db-collection@0.2.2
+  - @tanstack/query-db-collection@1.0.1
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`a83a818`](https://github.com/TanStack/db/commit/a83a8189514d22ca2fcdf34b9cb97206d3c03c38), [`a83a818`](https://github.com/TanStack/db/commit/a83a8189514d22ca2fcdf34b9cb97206d3c03c38)]:
+  - @tanstack/query-db-collection@1.0.1
+  - @tanstack/db@0.5.1
+  - @tanstack/electric-db-collection@0.2.1
+
 ## 0.0.2
 
 ### Patch Changes